# api/api-spec.yaml
openapi: 3.0.3
info:
  title: myOS API
  description: API for power users to interact with myOS programmatically
  version: 1.0.0
  contact:
    name: myOS Support
    email: support@myos.ai
servers:
  - url: https://api.myos.ai/v1
    description: Production server
  - url: https://staging-api.myos.ai/v1
    description: Staging server
security:
  - ApiKeyAuth: []
paths:
...
  /files/{file_path}:
    get:
      summary: Read file content
      parameters:
        - name: file_path
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: File content
          content:
            text/plain:
              schema:
                type: string
    
    put:
      summary: Write file content
      parameters:
        - name: file_path
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          text/plain:
            schema:
              type: string
      responses:
        '200':
          description: File updated successfully
...
  /agents/{agent_id}:
    get:
      summary: Get agent details
      parameters:
        - name: agent_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Agent details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    Agent:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [running, completed, failed, stopped]
        model:
          type: string
        prompt:
          type: string
        created_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        result:
          type: string
...